Add files via upload
This is a "Talk to Speak" web application, designed to convert your text into spoke . Build using  HTML for structure, CSS for an attractive and JavaScript for seamless functionality .It can offers a modern and interactive experience.  In this application consists two buttons one is Listen  and another one is Stop. Listen button can Stop.